* feat: 보너스 번호 추가 및 등수 판단 기능 변경

* feat: 당첨통계 출력기능 변경

* review-feedback: LottoGameResult 내 통계 기능 리팩토링

* review-feedback: LottoRank 리팩토링

+++ src/main/java/lotto/dto/LottoGameResultDTO.java
@@ -7,15 +7,15 @@ import lotto.rank.LottoRank;
  * 당첨 통계와 수익률을 같이 전달하기 위한 DTO 클래스
  */
 public class LottoGameResultDTO {
-    private Map<LottoRank, Integer> statistics;
+    private Map<LottoRank, Long> statistics;
     private double yield;
 
-    public LottoGameResultDTO(Map<LottoRank, Integer> statistics, double yield) {
+    public LottoGameResultDTO(Map<LottoRank, Long> statistics, double yield) {
         this.statistics = statistics;
         this.yield = yield;
     }
 
-    public Map<LottoRank, Integer> getStatistics() {
+    public Map<LottoRank, Long> getStatistics() {
         return statistics;
     }
